# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ECOnexão API
  description: API de domínio do aplicativo ECOnexão para turismo sustentável e rotas
    ecológicas.
  version: 1.0.0
paths:
...
  /api/v1/places/photos/{token}/metadata:
    get:
      tags:
      - Google Place Photos
      summary: Get Place Photo Metadata
      description: Metadata is intentionally only available after a trusted server-side
        grant.
      operationId: get_place_photo_metadata_api_v1_places_photos__token__metadata_get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GooglePhotoMetadataEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/places/photos/{token}:
    get:
      tags:
      - Google Place Photos
      summary: Get Place Photo
      operationId: get_place_photo_api_v1_places_photos__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      - name: maxHeightPx
        in: query
        required: false
        schema:
          type: integer
          maximum: 4800
          minimum: 1
          default: 600
          title: Maxheightpx
      - name: maxWidthPx
        in: query
        required: false
        schema:
          type: integer
          maximum: 4800
          minimum: 1
          default: 800
          title: Maxwidthpx
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Foto indisponível
        '410':
          description: Foto expirada
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GooglePhotoAttributionSchema:
      properties:
        display_name:
          type: string
          title: Display Name
        uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Uri
      type: object
      required:
      - display_name
      title: GooglePhotoAttributionSchema
    GooglePhotoMetadataEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/GooglePhotoMetadataSchema'
      type: object
      required:
      - data
      title: GooglePhotoMetadataEnvelope
    GooglePhotoMetadataSchema:
      properties:
        proxy_url:
          type: string
          title: Proxy Url
        expires_at:
          type: integer
          title: Expires At
        width_px:
          type: integer
          title: Width Px
        height_px:
          type: integer
          title: Height Px
        author_attributions:
          items:
            $ref: '#/components/schemas/GooglePhotoAttributionSchema'
          type: array
          title: Author Attributions
        google_maps_uri:
          type: string
          title: Google Maps Uri
      type: object
      required:
      - proxy_url
      - expires_at
      - width_px
      - height_px
      - google_maps_uri
      title: GooglePhotoMetadataSchema
      description: Ephemeral photo metadata; deliberately excludes Google resource/photo
        URLs.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer